Revert "refactor(Dockerfile): remove useless mkdir in the volume"
This reverts commit 248f08e3c234539cc13580b59617cd3d47bee466.

It's better to create the mandatory directories inside the Dockerfile than
using a side script.
I said "mkdir are discarded because at build time the volume is not mounted. Thus mkdir /volume/something is lost." but i was wrong.
From the official documentation: "Changing the volume from within the Dockerfile: If any build steps change the data within the volume after it has been declared, those changes will be discarded."